Rename command files to end with command
