+++ src/commands/AdminCommandsHandler.ts
@@ -1,6 +1,6 @@
 import { BaseSubcommandsHandler } from "../baseCommand";
 import AdminQueueCommandsHandler from "./admin/AdminQueueCommandsHandler";
-import UpdateBotRolesCommand from "./admin/UpdateBotRoles";
+import UpdateBotRolesCommand from "./admin/UpdateBotRolesCommand";
 
 export default class AdminCommandsHandler extends BaseSubcommandsHandler {
     public static name = "admin";
