+++ src/commands/PingCommand.ts
