+++ src/commands/index.ts
@@ -1,5 +1,5 @@
-import PingCommand from './Ping';
-import HelpCommand from './Help';
+import PingCommand from './PingCommand';
+import HelpCommand from './HelpCommand';
 import AdminCommandsHandler from './AdminCommandsHandler';
 import ConfigCommandsHandler from './ConfigCommandsHandler';
