+++ tests/commands/admin/UpdateBotRoles.test.ts
@@ -1,6 +1,6 @@
 import { ApplicationCommandOptionType, BaseMessageOptions, ChatInputCommandInteraction, EmbedBuilder, PermissionsBitField, Role, RoleCreateOptions } from "discord.js";
 import { MockDiscord } from "../../mockDiscord";
-import UpdateBotRolesCommand from "../../../src/commands/admin/UpdateBotRoles";
+import UpdateBotRolesCommand from "../../../src/commands/admin/UpdateBotRolesCommand";
 import { container } from "tsyringe";
 import { mockRole } from "@shoginn/discordjs-mock";
 import { InternalGuildRoles } from "../../../src/models/BotRoles";
